fix: prevent having to pass renderer to all trees
